+++ Projekt pk4/space invaders ++/MediumEnemy.cpp
@@ -37,14 +37,11 @@ void MediumEnemy::shoot()
 		Clock.restart();
 		if (Xshootingpatern)
 		{
-			bullets.push_back(new Bullet("MediumBullet",this->position, Vector2i(0.70710678118*bulletspeed, 0.70710678118*bulletspeed),150));
-			bullets.push_back(new Bullet("MediumBullet",this->position, Vector2i(-0.70710678118*bulletspeed, 0.70710678118*bulletspeed),150));
-			bullets.push_back(new Bullet("MediumBullet",this->position, Vector2i(0.70710678118*bulletspeed, -0.70710678118*bulletspeed),150));
-			bullets.push_back(new Bullet("MediumBullet",this->position, Vector2i(-0.70710678118*bulletspeed, -0.70710678118*bulletspeed),150));
-//			bullets.push_back(new Bullet(this->position, bulletspeed, 0.75*Pi));
-//			bullets.push_back(new Bullet(this->position, bulletspeed, 0.25*Pi));
-//			bullets.push_back(new Bullet(this->position, bulletspeed, 1.25*Pi));
-//			bullets.push_back(new Bullet(this->position, bulletspeed, 1.75*Pi));
+			bullets.push_back(new Bullet("MediumBullet",this->position, Vector2i((int)(0.70710678118*(double)bulletspeed), (int)(0.70710678118*(double)bulletspeed)),150));
+			bullets.push_back(new Bullet("MediumBullet",this->position, Vector2i((int)(-0.70710678118*(double)bulletspeed),(int)( 0.70710678118*(double)bulletspeed)),150));
+			bullets.push_back(new Bullet("MediumBullet",this->position, Vector2i((int)(0.70710678118*(double)bulletspeed), (int)(-0.70710678118*(double)bulletspeed)),150));
+			bullets.push_back(new Bullet("MediumBullet",this->position, Vector2i((int)(-0.70710678118*(double)bulletspeed),(int)( -0.70710678118*(double)bulletspeed)),150));
+
 			Xshootingpatern = 0;
 		}
 		else
@@ -53,10 +50,7 @@ void MediumEnemy::shoot()
 			bullets.push_back(new Bullet("MediumBullet",this->position, Vector2i(0,-bulletspeed),150));
 			bullets.push_back(new Bullet("MediumBullet",this->position, Vector2i(bulletspeed,0),150));
 			bullets.push_back(new Bullet("MediumBullet",this->position, Vector2i(-bulletspeed,0),150));
-		//	bullets.push_back(new Bullet(this->position, bulletspeed, 0));
-		//	bullets.push_back(new Bullet(this->position, bulletspeed, 0.5*Pi));
-		//	bullets.push_back(new Bullet(this->position, bulletspeed, Pi));
-		//	bullets.push_back(new Bullet(this->position, bulletspeed, 1.5*Pi));
+
 			Xshootingpatern = 1;
 		}
 	}
