+++ Projekt pk4/space invaders ++/SmallEnemy.cpp
@@ -56,7 +56,7 @@ void SmallEnemy::shoot()
 	{
 		Clock.restart();
 		bullets.push_back(new Bullet("SmallBullet",Vector2u(this->position.x, this->position.y + (GetSize().y / 2)),Vector2i(0, bulletspeed),100));
-		bullets.push_back(new Bullet("SmallBullet",Vector2u(this->position.x, this->position.y + (GetSize().y / 2)), Vector2i((int)(bulletspeed*sin30deg),bulletspeed*cos30deg),100));
-		bullets.push_back(new Bullet("SmallBullet",Vector2u(this->position.x, this->position.y + (GetSize().y / 2)), Vector2i((int)(bulletspeed*-sin30deg), bulletspeed*cos30deg), 100));
+		bullets.push_back(new Bullet("SmallBullet",Vector2u(this->position.x, this->position.y + (GetSize().y / 2)), Vector2i((int)(bulletspeed*sin30deg),(int)(bulletspeed*cos30deg)),100));
+		bullets.push_back(new Bullet("SmallBullet",Vector2u(this->position.x, this->position.y + (GetSize().y / 2)), Vector2i((int)(bulletspeed*-sin30deg), (int)(bulletspeed*cos30deg)), 100));
 	}
 }
